added the requied icons for login page and also moved the login page to its directory, also added its router in app.jsx
