+++ src/Pages/Public/Login/LoginSignup.jsx
@@ -1,0 +1,63 @@
+import React, { useState } from "react";
+import "./LoginSignup.css";
+import user_icon from "../assets/person.png";
+import email_icon from "../assets/email.png";
+import password_icon from "../assets/password.png";
+
+const LoginSignup = () => {
+  const [action, setAction] = useState("Login");
+
+  return (
+    <div className="container">
+      <div className="header">
+        <div className="text">{action}</div>
+        <div className="underline"></div>
+      </div>
+
+      <div className="inputs">
+        {/* Show Name field only when Sign Up is selected */}
+        {action === "Sign Up" && (
+          <div className="input">
+            <img src={user_icon} alt="User Icon" />
+            <input type="text" placeholder="Name" />
+          </div>
+        )}
+
+        <div className="input">
+          <img src={email_icon} alt="Email Icon" />
+          <input type="email" placeholder="Email Id" />
+        </div>
+
+        <div className="input">
+          <img src={password_icon} alt="Password Icon" />
+          <input type="password" placeholder="Password" />
+        </div>
+      </div>
+
+      {/* Show Forgot Password only for Login */}
+      {action === "Login" && (
+        <div className="forgot-password">
+          Lost password? <span>Click Here</span>
+        </div>
+      )}
+
+      <div className="submit-container">
+        <button
+          className={action === "Login" ? "submit gray" : "submit"}
+          onClick={() => setAction("Sign Up")}
+        >
+          Sign Up
+        </button>
+
+        <button
+          className={action === "Sign Up" ? "submit gray" : "submit"}
+          onClick={() => setAction("Login")}
+        >
+          Login
+        </button>
+      </div>
+    </div>
+  );
+};
+
+export default LoginSignup;
